requirement.txt
pystreamer>=0.4.1
requests
pyyaml